requirements.txt
asgiref==3.7.2
astroid==3.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cloudinary==1.38.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
dill==0.3.8
dj-database-url==1.2.0
Django==5.0.2
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
psycopg==3.1.18
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isort==5.13.2
mccabe==0.7.0
Pillow==10.0.0
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pylint==3.0.3
pyOpenSSL==24.0.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
urllib3==2.2.1
zope.interface==6.2
